feat: add user project list

+++ internal/models/project_table.go
@@ -8,7 +8,7 @@ import (
 type Project struct {
 	gorm.Model
 	// Status 1: normal 2: ban | 状态 1 正常 2 禁用
-	ProjectStatus uint8 `gorm:"column:project_status" json:"project_status,omitempty"`
+	ProjectStatus uint32 `gorm:"column:project_status" json:"project_status,omitempty"`
 	// Project name | 角色名
 	ProjectName string `gorm:"column:project_name" json:"project_name,omitempty"`
 	// Remark | 备注
